--- src/components/SignUp/SignUpForm/index.tsx
@@ -1,5 +1,0 @@
-import * as React from "react";
-
-export default function SignUpForm() {
-  return <div>SignUpForm</div>;
-}
