--- src/components/SignUp/SignUpInput/index.tsx
@@ -1,5 +1,0 @@
-import * as React from "react";
-
-export default function SignUpInput() {
-  return <div>index</div>;
-}
