+++ src/components/common/Loading/index.tsx
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export default function Loading() {
+  return <div>Loading</div>;
+}
